requirements.txt
asgiref==3.8.1
boto3==1.35.41
botocore==1.35.41
Django==5.1.2
django-allauth==65.0.2
django-cors-headers==4.5.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
jmespath==1.0.1
pillow==11.0.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.3
